add helper to Post module , fetch and show news

+++ Modules/Post/Database/Migrations/2018_10_14_073331_create_news_rating_table.php
@@ -15,7 +15,7 @@ class CreateNewsRatingTable extends Migration
     {
         Schema::create('rates', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('rate');
+            $table->float('rate');
             $table->integer('post_id');
             $table->integer('user_id');
             $table->timestamps();
